Footer is refactored
